+++ app/src/ai/agent_sdk/setup_observability.rs
@@ -155,21 +155,8 @@ impl OzRunTimelineEvent {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::OzRunTimelineEvent;
-
-    #[test]
-    fn setup_timeline_event_names_match_server_contract() {
-        assert_eq!(
-            OzRunTimelineEvent::AgentStarted.as_event_name(),
-            "agent_started"
-        );
-        assert_eq!(
-            OzRunTimelineEvent::WorkerContainerReady.as_event_name(),
-            "worker_container_ready"
-        );
-    }
-}
+#[path = "setup_observability_tests.rs"]
+mod tests;
 
 #[derive(Clone, Copy)]
 pub(crate) enum SetupStep {
